pyproject.toml
[project]
name = "multilingual-literature-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "notebook>=7.4.4",
    "numpy>=2.2.6",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "pytesseract>=0.3.13",
    "regex>=2024.11.6",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.0.0",
    "tiktoken>=0.9.0",
    "uvicorn>=0.35.0",
]
